refactor(core): merge selector builder and rename polarity model
With the negated adapter gone, EntitySelectorAdapter and
EntitySelectorState had a single writer and a single reader each, so the
adapter/state split and its assignX indirection no longer paid rent.
Merge them into one EntitySelectorBuilder implementing the full scope;
the renderer stays separate as the serialization boundary.

Rename the internal polarity variants to SelectorFilter.Positive and
SelectorFilter.Negative so the public Excluded<T> call-site wrapper is
the only "excluded" concept, and move the !-prefix rendering out of the
model into EntitySelectorRenderer.
