+++ modules/core/src/main/kotlin/io/github/lmliam/kotventure/core/selector/SelectorFilter.kt
@@ -1,11 +1,14 @@
 package io.github.lmliam.kotventure.core.selector
 
+/**
+ * Polarity of a singleton selector argument: one positive value, or accumulated exclusions.
+ */
 internal sealed interface SelectorFilter<out T> {
-    data class Included<T>(
+    data class Positive<T>(
         val value: T,
     ) : SelectorFilter<T>
 
-    data class Excluded<T>(
+    data class Negative<T>(
         val values: List<T>,
     ) : SelectorFilter<T>
 }
@@ -15,10 +18,10 @@ internal fun <T> SelectorFilter<T>?.including(
     value: T,
 ): SelectorFilter<T> =
     when (this) {
-        null -> SelectorFilter.Included(value)
-        is SelectorFilter.Included ->
+        null -> SelectorFilter.Positive(value)
+        is SelectorFilter.Positive ->
             error("Selector argument '$argument' is already set; vanilla syntax allows one positive value.")
-        is SelectorFilter.Excluded ->
+        is SelectorFilter.Negative ->
             error(
                 "Selector argument '$argument' already has exclusions; vanilla syntax cannot combine them with a positive value.",
             )
@@ -29,16 +32,10 @@ internal fun <T> SelectorFilter<T>?.excluding(
     value: T,
 ): SelectorFilter<T> =
     when (this) {
-        null -> SelectorFilter.Excluded(listOf(value))
-        is SelectorFilter.Excluded -> copy(values = values + value)
-        is SelectorFilter.Included ->
+        null -> SelectorFilter.Negative(listOf(value))
+        is SelectorFilter.Negative -> copy(values = values + value)
+        is SelectorFilter.Positive ->
             error(
                 "Selector argument '$argument' already has a positive value; vanilla syntax cannot combine it with exclusions.",
             )
-    }
-
-internal fun <T> SelectorFilter<T>.renderValues(render: (T) -> String): List<String> =
-    when (this) {
-        is SelectorFilter.Included -> listOf(render(value))
-        is SelectorFilter.Excluded -> values.map { value -> "!${render(value)}" }
     }
